fix(ci): python-coverage-shards — split two files back under the 1500-line guardrail
PR #639's additions pushed two first-party files past the
`MAX_FIRST_PARTY_FILE_LINES = 1500` maintainability guardrail
(`tests/unit/test_core_decomposition_maintainability.py::
test_first_party_code_files_stay_under_line_limit`), which runs inside the
`python-coverage-shards` job and fails deterministically:

  - src/awf/service/orphan_resources.py  1560
  - tests/unit/service/test_worker.py    1516

Decompose both along existing seams rather than weakening the check:

  - Extract the reap-result dataclasses (`OrphanReapOutcome`,
    `OrphanReapResult`), the compose-teardown closure builder, and the
    pure row-less orphan aging/`--limit` helpers into a new
    `awf.service.orphan_reaping` module. `reap_classified_orphans` /
    `sweep_classified_orphans` stay in `orphan_resources` because their
    behavior is pinned to that module's monkeypatched scan/classify seam;
    they (and external callers) keep importing the moved names via an
    explicit re-export, so the public surface is unchanged. The new module
    depends on `orphan_resources` only through annotation-only
    (TYPE_CHECKING) types, so there is no runtime import cycle.

  - Move the self-contained `_service_git_environment` tests into a new
    `tests/unit/service/test_worker_git_environment.py` sibling.

orphan_resources.py -> 1393, test_worker.py -> 1304; both new files well
under the limit. No behavior change; full reaping/worker/mcp/status suites
and the guardrail pass.
